add relation user and notes

+++ src/note/note.service.ts
@@ -9,6 +9,7 @@ import { CreateNoteDto } from './dto/create-note.dto';
 import { UpdateNoteDto } from './dto/update-note.dto';
 import { Note } from './entities/note.entity';
 import { PaginationDto } from '../common/dtos/pagination.dto';
+import { User } from 'src/auth/entities/user.entity';
 
 @Injectable()
 export class NoteService {
@@ -17,9 +18,10 @@ export class NoteService {
     private readonly noteRepository: Repository<Note>,
   ) {}
 
-  async create(createNoteDto: CreateNoteDto) {
+  async create(createNoteDto: CreateNoteDto, user: User) {
     try {
       const note = this.noteRepository.create(createNoteDto);
+      note.user = user;
       await this.noteRepository.save(note);
       return note;
     } catch (error) {
@@ -27,27 +29,42 @@ export class NoteService {
     }
   }
 
-  async findAll(PaginationDto: PaginationDto) {
+  async findAll(PaginationDto: PaginationDto, user: User) {
     const { itemPerPage, currentPage, keyword } = PaginationDto;
-    console.log('item per page', itemPerPage);
-    console.log('current page', currentPage);
-    console.log('keyword', keyword);
+
     try {
-      const [notes, total] = await this.noteRepository.findAndCount({
-        where: { title: Like(`%${keyword}%`) },
-        take: itemPerPage,
-        skip: (currentPage - 1) * itemPerPage,
-      });
-      return { currentPage, itemPerPage, total, notes };
+      if (keyword) {
+        const [notes, total] = await this.noteRepository.findAndCount({
+          where: {
+            user,
+            title: Like(`%${keyword}%`),
+          },
+          skip: (currentPage - 1) * itemPerPage,
+          take: itemPerPage,
+          order: { id: 'DESC' },
+        });
+        return { currentPage, itemPerPage, total, notes };
+      } else {
+        const [notes, total] = await this.noteRepository.findAndCount({
+          where: { user },
+          skip: (currentPage - 1) * itemPerPage,
+          take: itemPerPage,
+          order: { id: 'DESC' },
+        });
+        return { currentPage, itemPerPage, total, notes };
+      }
     } catch (error) {
       this.handlerException(error);
     }
   }
 
-  async findOne(id: number) {
+  async findOne(id: number, user: User) {
     try {
       const note = await this.noteRepository.find({
-        where: { id },
+        where: {
+          id,
+          user,
+        },
         select: ['id', 'title', 'content'],
       });
       return note;
@@ -56,9 +73,9 @@ export class NoteService {
     }
   }
 
-  async update(id: number, updateNoteDto: UpdateNoteDto) {
+  async update(id: number, updateNoteDto: UpdateNoteDto, user: User) {
     try {
-      const note = await this.noteRepository.findOneBy({ id: id });
+      const note = await this.noteRepository.findOneBy({ id: id, user: user });
       if (!note) {
         throw new BadRequestException('Note not found');
       }
@@ -69,8 +86,17 @@ export class NoteService {
     }
   }
 
-  remove(id: number) {
-    this.noteRepository.delete(id);
+  async remove(id: number, user: User) {
+    console.log({ id, user });
+
+    //check id note exist and belong to user
+    const note = await this.noteRepository.findOneBy({ id: id, user: user });
+    if (!note) {
+      throw new BadRequestException('Note not found');
+    }
+    console.log({ note });
+
+    return this.noteRepository.delete(id);
   }
   private handlerException(error: any) {
     if (error.code === 11000) {
